# [#103] feat: Add GEMINI.md documentation and SqlExpiredEvents example