… GEMINI.md
… src/main/java/com/yugabyte/sample/apps/SqlExpiredEvents.java