Add `GEMINI.md` to serve as a guide for creating new applications and workload generators with the Gemini CLI.

Outline the required steps for using the CLI:

1. Generate a new application from workload and schema requirements.
2. Format the generated code to match the YugabyteDB code style.
3. Prompt the user for the information needed to build and run the app.
4. Record the original prompt to ensure reproducibility.

Add the `SqlExpiredEvents` application to demonstrate this workflow. The application must show how index conditions are not pushed down effectively when using a `timestamp` column. Provide a flag to toggle between `timestamp` and `timestamptz` to illustrate the performance difference.

The resulting performance comparison is as follows:

* Without --use_timestamptz: ~0.36 ops/sec | ~2749.72 ms/op
* With    --use_timestamptz: ~3.72 ops/sec |  ~269.93 ms/op

The 10x difference is explained by the fact that the index scan only matches 1/10th of the rows.